backend/requirements.txt
flask
flask-cors
opencv-python
ultralytics
requests
python-dotenv
